backend/requirements_django.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
django-multitenant>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
bcrypt>=4.2
gunicorn>=22.0
